Add files via upload
One of my notable projects is an interior decor website that showcases my ability to blend aesthetic design with functional coding. This project demonstrates my attention to detail, creativity, and technical expertise in delivering a user-friendly platform for interior design enthusiasts. It's built with HTML, CSS, Javascript, MySQL, and PHP.